Refactored tests to work with Java 11
